Updated the hw for lesson 5. All works correctly. But there is a question "how to replace Student.class with T.class?"(see TODOs in the AbstractEntityDao).
